refactor(server): extract startup preflight stage
Keep the non-blocking Codex hook trust-hash probe as an explicit bootstrap stage so later server::run decomposition can compose it without changing boot order.
